+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for NeoFS Monitor
 ## [Unreleased]
 
 ### Added
+- Node capacity and overall cluster size metric (#177)
 
 ### Changed
 - github.com/nspcc-dev/neofs-sdk-go v1.0.0-rc.14 => v1.0.0-rc.15.0.20251014091731-abf7dc3b4b64 (#177)